change migration currency rate

+++ database/migrations/2021_12_25_173036_create_currency_rate_table.php
@@ -17,7 +17,7 @@ class CreateCurrencyRateTable extends Migration
             $table->id();
             $table->string('currency_code_from_id');
             $table->string('currency_code_to_id');
-            $table->integer('value')->nullable();
+            $table->double('value')->nullable();
             $table->boolean('is_active')->default(false);
             $table->timestamps();
             $table->softDeletes();
